gcp-serverless-app/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-storage
